feat(dupreq): don't retire inflight entries
Currently, entries in the fifo queue can be retired no matter
whether they are inflight or completed. If the inflight entry is
retired and then the retry request comes in, another DRC entry
will be allocated to this request and it will be also dispatched
and processed, at the same time with the first one. The same two
inflight requests may lead to data inconsistenty including the
"lost write due to duplicated truncate" issue.

Considering the following case:
1. The application wants to truncate a existent file and write it.
It opens the file and sends SETATTR request.
2. FSAL is not available to handle SETATTR and hangs there.
3. There are many other requests processed successfully and the
first SETATTR DRC entry is retired.
4. Timeout arrives and NFS client retry SETATTR request.
5. The second SETATTR is dispatched into FSAL and also hangs.
6. The first SETATTR completes successfully and send reply to
NFS client.
7. NFS client sends WRITE requests.
8. WRITE completes successfully.
9. The second SETATTR finishes and reset the WRITE in step 8,
which means the write is lost.

This commit only retires completed entries and always keep the
inflight entry in DRC to suspend retried requests. In the above
example, the inflight SETATTR entry is NOT retired in step 3,
and the second SETATTR will be suspend in step 5, and it will be
ignored in step 6. So the write will not be lost.

See more:
https://github.com/nfs-ganesha/nfs-ganesha/issues/1253

This commit also brings other changes:
1. set dupreq_entry_t.rc on nfs_dupreq_finish
2. acquire partition lock firstly on delete()
3. set default DRC_TCP_Hiwat to half of DRC_TCP_Size
4. update some comments

Change-Id: I721ef5ff934362ae63624709041a19b268581173
Signed-off-by: zhitaoli <zhitao.li@iomesh.com>